pyproject.toml
[tool.poetry]
name = "retail-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "retail_api"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
